split distill worker
